requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
redis==5.0.4
aiodns==3.1.1
dnspython==2.6.1
dkimpy==1.1.5
python-spf==0.0.1